fix(runtime): internal: issues accessing resources
Fixes issues when accessing resources in compose desktop and AOSP
to grab the signatures of annotated elements.

Comment:

Although the JVM is a revolutionary concept,
implementations of JRE and the different
environments and setups can cause some features
to be absolutely missed up.

In this fix, I tried to savage all information
and details about these setups and here is a brief
of the problems I solved in this commit.

android: AOSP does not implement resource directory
  locating when using ClassLoader.getResources() or
  any similar functions. This is a huge problem for
  ranno since it depends on enumerating the items of
  a specific resource directory for it to read the
  signatures of annotated elements.

compose-desktop: when using the createDistributable gradle
  task, the output includes an embed jre that has some
  functionality missing, in my case, the functionality was
  ZipFileSystemProvider which was necessary for enumerating
  items in a directory which is a deal-breaker for ranno.

finally, the solutions that was implemented are not THE solution
for the original problem (jvm resources is not unified experience)
and future issues might come. Additionally, the solutions here are
actually workarounds that satisfies the needs of ranno and highly
doubt will work on any other library or needs.